+++ src/BookNavigator/book-navigator.js
